Encapsulado los dropdowns
*Se mueven los dropdowns a un componente DropdownFilter para futuro uso
